requirements/requirements.txt
python-dotenv
pandas
pymysql
sqlalchemy
requests
cryptography
peewee
requests